Week_3/03102025/accounts_service/requirements.txt
fastapi>=0.112.0
uvicorn>=0.30.0
SQLAlchemy>=2.0.31
pydantic>=2.7.0
pydantic-settings>=2.3.0
python-dotenv>=1.0.1
httpx>=0.27.0
pytest>=8.2.0
pydantic[email]
